Factor LiveDataAdapter across files
